+++ src/components/login/login.css
@@ -1,0 +1,15 @@
+.login__box{
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: 2rem 0;
+}
+
+.login__form{
+    padding: 2rem;
+    background-color: var(--color-primary);
+
+    display: flex;
+    flex-direction: column;
+}
